+++ include/pbvp/hook_manager.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-namespace pbvp::hooks {
-
-bool ProbeAndInstall() noexcept;
-void MarkShutdown() noexcept;
-bool IsReady() noexcept;
-
-} // namespace pbvp::hooks